% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pink.R
\name{pink}
\alias{pink}
\title{Simulate response data with pink noise profile}
\usage{
pink(N, alpha = 1)
}
\arguments{
\item{N}{Number of trials.}

\item{alpha}{Value of alpha (default 1).}
}
\value{
A vector with response values.
}
\description{
Simulate response data with pink noise profile
}
