% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{simdata_1f}
\alias{simdata_1f}
\title{Simulate data from a one-factor design}
\usage{
simdata_1f(
  etastep,
  nPart,
  nLevels,
  nReps,
  nSubblocks,
  mu = 0,
  etaSQ = etaSQ_1f(nPart, nLevels, nReps)[etastep],
  prop_rint = 0.35,
  prop_rslp = 0.11,
  errstr = c("noauto", "expDecay", "pinkNoise", "randomWalk", "mixed")
)
}
\arguments{
\item{etastep}{Step number to determine eta^2 value (1-7).}

\item{nPart}{Number of participants (25, 40, or 60).}

\item{nLevels}{Number of levels of the independent variable (2, 4, 8).}

\item{nReps}{Number of repetitions within each level for each participant (6, 12, 24).}

\item{nSubblocks}{Number of subblocks.}

\item{mu}{Intercept.}

\item{etaSQ}{Eta-squared (measure of the effect size).}

\item{prop_rint}{Proportion of variance accounted for by random intercepts.}

\item{prop_rslp}{Proportion of variance accounted for by random slopes.}

\item{errstr}{Error structure ("noauto", "expDecay", "pinkNoise", "randomWalk", or "mixed")}
}
\value{
A data frame with simulated data.
}
\description{
Simulate data from a one-factor design
}
