% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\name{simdata_2x2ww}
\alias{simdata_2x2ww}
\title{Simulate data from a 2x2 factorial design}
\usage{
simdata_2x2ww(
  etastep,
  nPart,
  nReps,
  nSubblocks,
  mu = 0,
  etaSQ_row = etaSQ_2x2ww(nPart, nReps)[etastep],
  etaSQ_col = etaSQ_2x2ww(nPart, nReps)[etastep],
  etaSQ_rcx = etaSQ_2x2ww(nPart, nReps)[etastep],
  prop_rint = 0.11,
  prop_rslp_row = 0.11,
  prop_rslp_col = 0.11,
  prop_rslp_rcx = 0.11,
  errstr = c("noauto", "expDecay", "pinkNoise", "randomWalk", "mixed")
)
}
\arguments{
\item{etastep}{Step number to determine eta^2 value (1-7).}

\item{nPart}{Number of participants (28, 40, or 50).}

\item{nReps}{Number of repetitions within each level for each participant (10, 16).}

\item{nSubblocks}{Number of subblocks.}

\item{mu}{Intercept.}

\item{etaSQ_row}{Eta-squared (measure of the effect size) for the row factor ("A").}

\item{etaSQ_col}{Eta-squared for the column factor ("B").}

\item{etaSQ_rcx}{Eta-squared for the interaction.}

\item{prop_rint}{Proportion of variance accounted for by random intercepts.}

\item{prop_rslp_row}{Proportion of variance accounted for by random slopes for row.}

\item{prop_rslp_col}{Proportion of variance accounted for by random slopes for column.}

\item{prop_rslp_rcx}{Proportion of variance accounted for by random slopes for interaction.}

\item{errstr}{Error structure ("noauto", "expDecay", "pinkNoise", "randomWalk", or "mixed")}
}
\value{
A data frame with simulated data.
}
\description{
Simulate data from a 2x2 factorial design
}
